requirements.txt
python-dotenv
whiskeysockets
supabase
asyncio
requests
